tests/lib/hardware: Restructure hardware related code
Move the base class to its own file and the OpenStack libcloud
implementation to it's own file. That way, it's easier to write new
hardware backends.
